Experiment editing happening and trial adding done

+++ Project/LabNet/app/src/main/java/com/example/lab_net/BinomialTrial.java
@@ -5,7 +5,7 @@ public class BinomialTrial extends Trial{
     private int fail;
 
 
-    public BinomialTrial(String id, String title, int result) {
+    public BinomialTrial(String id, String title, Long result) {
         super(id, title, result);
     }
 }
