+++ Project/LabNet/app/src/main/java/com/example/lab_net/CountTrial.java
@@ -4,7 +4,7 @@ public class CountTrial extends Trial{
     private int count;
 
 
-    public CountTrial(String id, String title, int result) {
+    public CountTrial(String id, String title, Long result) {
         super(id, title, result);
     }
 }
