+++ Project/LabNet/app/src/main/java/com/example/lab_net/NonNegativeIntegerTrial.java
@@ -4,7 +4,7 @@ public class NonNegativeIntegerTrial extends Trial{
     private int nonNegativeCount;
 
 
-    public NonNegativeIntegerTrial(String id, String title, int result) {
+    public NonNegativeIntegerTrial(String id, String title, Long result) {
         super(id, title, result);
     }
 }
